SCRUM-11 1. Corrected the data type in OrderStatusScheduler and OrderRepository.

+++ src/main/java/com/example/finalproject/repository/OrderRepository.java
@@ -22,8 +22,8 @@ public interface OrderRepository extends JpaRepository<Order, Long> {
             "o.updatedAt = CURRENT_TIMESTAMP " +
             "WHERE o.createdAt > :startDate " +
             "AND o.status IN (:statuses)")
-    long updateOrderStatuses(Timestamp startDate, @Param("statuses") List<Status> statuses);
+    Integer updateOrderStatuses(Timestamp startDate, @Param("statuses") List<Status> statuses);
 
     @Query("SELECT COUNT(o) FROM Order o WHERE o.status != 'DELIVERED' AND o.status != 'CANCELED' AND o.createdAt > :startDate")
-    long countUndeliveredOrders(Timestamp startDate);
+    Integer countUndeliveredOrders(Timestamp startDate);
 }
